refactor: refactor ChallengeCategories
